cloud_functions/delete_data/requirements.txt
Flask==2.0.0
google-cloud-secret-manager==2.7.2
google-cloud-core==1.7.1
google-auth==1.33.1
psycopg2==2.9.1